requirements.txt
flask>=2.3.0
flask-login>=0.6.3
flask-wtf>=1.1.1
werkzeug>=2.3.0
bcrypt>=4.0.1
APScheduler>=3.10.4
PyGithub>=1.59.1
gitpython>=3.1.32
python-dotenv>=1.0.0
requests>=2.31.0
sqlalchemy>=2.0.21
flask-sqlalchemy>=3.0.5
wtforms>=3.0.1
gunicorn>=21.2.0
